requirements.txt
numpy==1.24.3
openai-whisper==20230314
Wave==0.0.2
openai==0.27.6
customtkinter==5.1.3
PyAudioWPatch==0.2.12.5
pyinstaller==5.13.0
--extra-index-url https://download.pytorch.org/whl/cu117
torch
pyperclip
PyYAML
